# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a API-Jira server.",
        "version": "0.0.1",
        "title": "Jira-API",
        "termsOfService": ""
    },
    "host": "",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/get/jira/issue/issueTypes": {
            "get": {
                "tags": [
                    "Get Jira Issue"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "projectIdOrKey",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "AM"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/issue/issueTypeNametoId": {
            "get": {
                "tags": [
                    "Get Jira Issue"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "projectIdOrKey",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "AM"
                    },
                    {
                        "name": "issueTypeName",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "Computer Asset"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/issue/allFields": {
            "get": {
                "tags": [
                    "Get Jira Issue"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/issue/issue": {
            "get": {
                "tags": [
                    "Get Jira Issue"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "issueId",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "TOCP-85"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/issue/issueNames": {
            "get": {
                "tags": [
                    "Get Jira Issue"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "issueId",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "IBVSD-23"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/issue/issueFields": {
            "get": {
                "tags": [
                    "Get Jira Issue"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "projectIdOrKey",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "AM"
                    },
                    {
                        "name": "issueTypeId",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "10502"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}